build: FreeBSD/macOS source-tree hierarchy + kernel -Werror bug-class hardening
Source tree
-----------
Complete the bin/ -> {bin,sbin,usr.bin,usr.sbin,tests} split (the file moves
landed separately; this wires up the build + path references):

  bin/      essentials needed early by the kernel/shell (init, login, shell,
            tcsh, ls, cp, cat, echo, ln, mkdir, mv, rm, sleep, date, ed, expr, pwd)
  sbin/     disk/system admin -> /sbin (fdisk, disklabel, format, mount,
            ubpool, ubfs, ubdisk, ping)
  usr.bin/  bulk user commands + GUI apps -> /usr/bin
  usr.sbin/ daemons/services -> /usr/sbin (views, vlogin, authd, logd,
            automountd, ubistry, aural, netcfg, dropbear)
  tests/    test/dev harnesses -> /usr/tests

Mechanism: each tree's Makefile.incl sets BINDIR; the shared
ubix.musl*.prog.mk emits the binary to ${OBJ_DIR}/${BINDIR}/ (and mkdir's it
in the link rule); the top Makefile's `world` descends every tree (Step 3b)
and `clean` follows.  mkimage.sh stages /usr/tests and applies the base
profile filter to the /usr copy loop (so a base image still skips
views/vlogin/settings now that they live under /usr).  gen-compile-commands.py
walks the new trees.

Path references updated so the moves don't regress boot: etc/init.d/* daemons
exec /usr/sbin/X; tools/makereg.c start-menu seeds point at /usr/bin|/usr/sbin;
bin/init/main.c CONSOLE_DESKTOP="/usr/sbin/views" and PATH gains /usr/tests;
the views suite's internal launch paths (vlogin, taskbar, term, settings,
netcfg) and vdoom's shared-DOOM-source path follow the moves.  Docs (README,
CLAUDE.md, ARCHITECTURE.md, BUILDING.md, filesystem-hierarchy-plan) describe
the new layout.

Kernel -Werror hardening
------------------------
Promote a narrow set of "almost always a real bug" warning classes to errors
on first-party kernel code (WERROR_BUGCLASSES): implicit-function-declaration,
implicit-int, int-to-pointer-cast, pointer-to-int-cast, return-type — the LP64
/ ABI families uBixOS has actually been bitten by.  This surfaced a real bug:

  devfs stashed a 64-bit device pointer in the 32-bit fd->start and read it
  back as a pointer (truncated on LP64; worked only by low-memory luck).  Moved
  it to the existing 64-bit-safe fd->res (the field ramfs already uses for
  this), updating devfs + the stat(2) devfs-root sentinel check.

Five benign/dead/debug narrowings (dma_paddr, mprotect addr, signal_check
debug print, ubthread cond/mutex id, ubthread_create) are made explicit via
(uintptr_t) routing — no behaviour change.  Both kernels build clean.

+++ sys/include/sys/dma_mem.h
@@ -49,7 +49,7 @@ void	 dma_free(struct dma_buf *buf);
 static inline u_int32_t
 dma_paddr(void *vaddr)
 {
-	return ((u_int32_t)vaddr);
+	return ((u_int32_t)(uintptr_t)vaddr);
 }
 
 #endif /* _SYS_DMA_MEM_H */
